+++ src/api/router.go
@@ -61,7 +61,7 @@ func newAPIs(b *bll.Blls, s *service.Services) *APIs {
 }
 
 func newRouters(apis *APIs) []*gear.Router {
-	return []*gear.Router{newRouterV1(apis), newRouterAPIV1(apis)}
+	return []*gear.Router{newRouterV1(apis), newRouterAPIV1(apis), newRouterServiceAPI(apis)}
 }
 
 func newRouterV1(apis *APIs) *gear.Router {
@@ -79,6 +79,17 @@ func newRouterV1(apis *APIs) *gear.Router {
 	return routerV1
 }
 
+func newRouterServiceAPI(apis *APIs) *gear.Router {
+	routerV2 := gear.NewRouter(gear.RouterOptions{
+		Root: "/api/v1",
+	})
+	routerV2.Use(tracing.New("urbs-console"))
+	routerV2.Use(middleware.VerifyService(services))
+	// 读取指定用户的功能配置项，包含 group，支持条件筛选，数据用于服务端
+	routerV2.Get("/users/:uid/settings:unionAll", apis.User.ListSettingsUnionAllBackend)
+	return routerV2
+}
+
 func newRouterAPIV1(apis *APIs) *gear.Router {
 
 	routerV1 := gear.NewRouter(gear.RouterOptions{
@@ -90,8 +101,6 @@ func newRouterAPIV1(apis *APIs) *gear.Router {
 	// ***** client ******
 	// 读取指定用户的功能配置项，包含 group，支持条件筛选，数据用于客户端
 	routerV1.Get("/users/settings:unionAll", apis.User.ListSettingsUnionAllClient)
-	// 读取指定用户的功能配置项，包含 group，支持条件筛选，数据用于服务端
-	routerV1.Get("/users/:uid/settings:unionAll", apis.User.ListSettingsUnionAllBackend)
 
 	// 读取指定用户的标签，支持条件筛选，数据用于客户端
 	routerV1.Get("/users/labels", apis.User.ListLablesForClient)